requirements.txt
astroid==3.3.10
black==25.1.0
certifi==2025.6.15
charset-normalizer==3.4.2
click==8.2.1
dill==0.4.0
idna==3.10
isort==6.0.1
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
pylint==3.3.7
python-dotenv==1.1.1
requests==2.32.4
tomlkit==0.13.3
urllib3==2.5.0
